Edited Score
Edited issues with score that related to score multiplier and added a scoreCalculation value that allows scores to increase based off of a double variable.

+++ Assets/Scripts/UIVariables.cs
@@ -10,7 +10,7 @@ public class UIVariables : MonoBehaviour
     [SerializeField] private GameObject gameRun;
     private int score;
     private string scoreUI;
-    private long totalScore;
+    private double totalScore;
 
     
     void Start()
